requirements.txt
# Attribution Analytics - Python Dependencies
# Python 3.10+

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Snowflake
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# dbt
dbt-core==1.7.4
dbt-snowflake==1.7.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.12.1
isort==5.13.2
sqlfluff==3.0.0

# Utilities
faker==22.0.0

